renames folders to follow unitree name
